requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.2
python-multipart==0.0.12
PyPDF2==3.0.1
pdf2image==1.17.0
pillow>=10.4.0
python-magic==0.4.27
requests==2.32.3
openai>=1.54.0
tqdm==4.67.1
pandas==2.2.3
python-dotenv==1.0.1
PyMuPDF==1.23.18
elevenlabs==2.5.0
